+++ applications/lethe-rpt-cell-reconstruction-3d/rpt_cell_reconstruction_3d.cc
@@ -1,6 +1,8 @@
 // SPDX-FileCopyrightText: Copyright (c) 2021, 2024 The Lethe Authors
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception OR LGPL-2.1-or-later
 
+#include <core/utilities.h>
+
 #include <rpt/rpt_calculating_parameters.h>
 #include <rpt/rpt_cell_reconstruction.h>
 
@@ -14,21 +16,43 @@ main(int argc, char *argv[])
 {
   try
     {
-      if (argc != 2)
+      Utilities::MPI::MPI_InitFinalize mpi_initialization(argc, argv, 1);
+
+      ConditionalOStream pcout(
+        std::cout, (Utilities::MPI::this_mpi_process(MPI_COMM_WORLD) == 0));
+
+      auto [options, args] = parse_args(argc, argv);
+
+      // Print version information
+      if (options["-V"])
         {
-          std::cout << "Usage:" << argv[0] << " input_file" << std::endl;
-          std::exit(1);
+          pcout << "Running: " << concatenate_strings(argc, argv) << std::endl;
+
+          if (Utilities::MPI::this_mpi_process(MPI_COMM_WORLD) == 0)
+            print_version_info(pcout);
+
+          return EXIT_SUCCESS;
         }
 
-      Utilities::MPI::MPI_InitFinalize mpi_initialization(argc, argv, 1);
+      if (args.empty())
+        {
+          pcout << "Usage: " << argv[0] << " input_file" << std::endl;
+          return EXIT_FAILURE;
+        }
+
+      const std::string file_name(args[0]);
 
       ParameterHandler         prm;
       RPTCalculatingParameters rpt_parameters;
       rpt_parameters.declare(prm);
 
       // Parsing of the file
-      prm.parse_input(argv[1]);
+      prm.parse_input(file_name);
       rpt_parameters.parse(prm);
+
+      // Print parameters if needed
+      if (Utilities::MPI::this_mpi_process(MPI_COMM_WORLD) == 0)
+        print_parameters_to_output_file(pcout, prm, file_name);
 
       RPTCellReconstruction<3> rpt_cell_reconstruction(
         rpt_parameters.rpt_param,